Cleaned up some null check and error throwing.
Implemented the interface correctly to handle DTO null in one of the services where it mismatched the interface.

Added dapr.yaml changes and kestrel to start up the project in development mode from the yaml file. This allows us to use the Dapr Dashboard

Added TODO: Statements for 2 methods I noted that could be optimised. Please test and revert in a new PR if applicable and then also remove the TODO:

Removed MAUI from standard solution, it's still there, just not in solution explorer. A seperate .SLN should be created for this